Change speeds to float

+++ mbed/arm/main.cpp
@@ -37,7 +37,7 @@ MEMS thermo_sensors[] = { MEMS(p9, p10),
 
 int minima[] = {100, 172, 360};
 int maxima[] = {720, 334, 360};
-int speeds[] = {0.1, 0.2, 0.2};
+float speeds[] = {0.1, 0.2, 0.2};
 int inits[] = {maxima[0], maxima[1], 0};
 int goals[] = {maxima[0], maxima[1], 0};
